requirements/base.txt
# Forensic Collector - Cross-platform base dependencies
# License: GPL-3.0 (due to pymobiledevice3 dependency)

# GUI
PyQt6>=6.6.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0

# WebSocket
websockets>=12.0

# System monitoring
psutil>=5.9.0

# Utilities
python-dateutil>=2.8.0

# Disk encryption (BitLocker + LUKS)
dissect.fve>=3.9
dissect.hypervisor>=3.14
